requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-WTF
Flask-Limiter
cryptography
ecdsa
pyotp
qrcode
Pillow
twilio
python-dotenv
pandas
scikit-learn
joblib
gunicorn